+++ tests/single_key.rs
@@ -8,31 +8,32 @@ use dusk_bls12_381::BlsScalar;
 use dusk_bytes::Serializable;
 use dusk_pki::{PublicKey, SecretKey};
 use dusk_schnorr::Signature;
+use ff::Field;
 use rand::rngs::StdRng;
 use rand::SeedableRng;
 
 #[test]
 fn sign_verify() {
-    let rng = &mut StdRng::seed_from_u64(2321u64);
+    let mut rng = StdRng::seed_from_u64(2321u64);
 
-    let sk = SecretKey::random(rng);
-    let message = BlsScalar::random(rng);
+    let sk = SecretKey::random(&mut rng);
+    let message = BlsScalar::random(&mut rng);
     let pk = PublicKey::from(&sk);
 
-    let sig = Signature::new(&sk, rng, message);
+    let sig = Signature::new(&sk, &mut rng, message);
 
     assert!(sig.verify(&pk, message));
 }
 
 #[test]
 fn test_wrong_keys() {
-    let rng = &mut StdRng::seed_from_u64(2321u64);
+    let mut rng = StdRng::seed_from_u64(2321u64);
 
-    let sk = SecretKey::random(rng);
-    let wrong_sk = SecretKey::random(rng);
-    let message = BlsScalar::random(rng);
+    let sk = SecretKey::random(&mut rng);
+    let wrong_sk = SecretKey::random(&mut rng);
+    let message = BlsScalar::random(&mut rng);
 
-    let sig = Signature::new(&sk, rng, message);
+    let sig = Signature::new(&sk, &mut rng, message);
 
     // Derive random public key
     let pk = PublicKey::from(&wrong_sk);
@@ -42,11 +43,11 @@ fn test_wrong_keys() {
 
 #[test]
 fn to_from_bytes() {
-    let rng = &mut StdRng::seed_from_u64(2321u64);
+    let mut rng = StdRng::seed_from_u64(2321u64);
 
-    let sk = SecretKey::random(rng);
-    let message = BlsScalar::random(rng);
+    let sk = SecretKey::random(&mut rng);
+    let message = BlsScalar::random(&mut rng);
 
-    let sig = Signature::new(&sk, rng, message);
+    let sig = Signature::new(&sk, &mut rng, message);
     assert_eq!(sig, Signature::from_bytes(&sig.to_bytes()).unwrap());
 }